refactor(channelPage): 채널페이지 레이아웃 커스텀 훅 분리 - 팔로우 핸들러 로직 분리 - 컴포넌트 이름 변경
