+++ src/app/(route)/channel/[uid]/layout.tsx
@@ -1,14 +1,13 @@
 'use client';
 
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import { useRouter, useParams, useSelectedLayoutSegment } from 'next/navigation';
 import Header from '@/app/_components/Header/Header.server';
 import NavBar from '@/app/(route)/(main)/_components/NavBar/NavBar.client';
 import useNavToggle from '@/app/_store/stores/main/useNavToggle.client';
-import { createClient } from '../../../_utils/supabase/client';
-import ChannelProfile from '@/app/(route)/channel/[uid]/components/ChannelProfile';
+import ChannelProfile from '@/app/(route)/channel/[uid]/components/ChannelProfile.server';
 import Footer from '@/app/_components/Footer/footer';
-import { useFollowAction } from '@/app/_store/queries/follow/mutation';
+import useChannelData from '@/app/_hooks/channel/useChannelData';
 
 interface RootLayoutProps {
   children: React.ReactNode;
@@ -16,102 +15,20 @@ interface RootLayoutProps {
 
 const RootLayout = ({ children }: RootLayoutProps) => {
   const { isOpen } = useNavToggle();
-  const [isClient, setIsClient] = useState(false);
-  const [userInfo, setUserInfo] = useState<{ nickname: string; channel_intro: string; img_url: string } | null>(null);
-  const [followerCount, setFollowerCount] = useState<number>(0);
-  const [isFollowing, setIsFollowing] = useState<boolean>(false);
-  const [loggedInUserId, setLoggedInUserId] = useState<string | null>(null);
-
   const { uid } = useParams();
   const router = useRouter();
   const activeSegment = useSelectedLayoutSegment();
-  const isCommunityActive = !activeSegment || ['post', 'comment', 'detail', 'edit'].includes(activeSegment || '');
+  const isCommunityActive =
+    !activeSegment || ['post', 'comment', 'detail', 'edit'].includes(activeSegment || '');
 
-  const supabase = createClient();
-  const { followMutate, unfollowMutate } = useFollowAction();
-
-  const fetchUserInfo = async () => {
-    if (!uid) return;
-    const { data, error } = await supabase
-      .from('users')
-      .select('nickname, channel_intro, profile_img')
-      .eq('id', uid)
-      .single();
-
-    if (error) {
-      console.error('사용자 정보 불러오기 오류', error);
-    } else {
-      setUserInfo({
-        nickname: data.nickname,
-        channel_intro: data.channel_intro,
-        img_url: data.profile_img || '',
-      });
-    }
-  };
-
-  const fetchLoggedInUser = async () => {
-    const {
-      data: { user },
-      error,
-    } = await supabase.auth.getUser();
-
-    if (error) {
-      console.error('로그인된 사용자 정보 가져오기 오류', error);
-    } else {
-      setLoggedInUserId(user?.id || null);
-    }
-  };
-
-  const fetchFollowerData = async () => {
-    if (!uid) return;
-    try {
-      const { count, error: countError } = await supabase
-        .from('follows')
-        .select('*', { count: 'exact' })
-        .eq('following_id', uid);
-
-      if (countError) {
-        throw new Error('팔로워 수 가져오기 오류: ' + countError.message);
-      }
-      setFollowerCount(count || 0);
-
-      if (loggedInUserId) {
-        const { data: followData, error: followError } = await supabase
-          .from('follows')
-          .select('*')
-          .eq('follower_id', loggedInUserId)
-          .eq('following_id', uid)
-          .single();
-
-        if (followError && followError.code !== 'PGRST116') {
-          throw new Error('팔로우 상태 확인 오류: ' + followError.message);
-        }
-        setIsFollowing(!!followData);
-      }
-    } catch (error) {
-      console.error(error);
-    }
-  };
-
-  useEffect(() => {
-    setIsClient(true);
-  }, []);
-
-  useEffect(() => {
-    if (isClient && uid) {
-      fetchUserInfo();
-    }
-  }, [isClient, uid]);
-
-  useEffect(() => {
-    fetchLoggedInUser();
-  }, []);
-
-  useEffect(() => {
-    if (uid) {
-      fetchFollowerData();
-    }
-  }, [uid, loggedInUserId]);
+  const {
+    userInfo,
+    loggedInUserId,
+    followerCount,
+    isFollowing,
+    handleFollow,
+    handleUnfollow,
+  } = useChannelData(uid as string);
 
   const handleTabClick = (path: string) => {
     router.push(path);
@@ -123,38 +40,6 @@ const RootLayout = ({ children }: RootLayoutProps) => {
 
   const handleChannelManagementClick = () => {
     if (uid) router.push(`/channel/${uid}/settings`);
-  };
-
-  const handleFollow = () => {
-    if (!uid) return;
-    followMutate(
-      { uid: Array.isArray(uid) ? uid[0] : uid, nickname: userInfo?.nickname || '' },
-      {
-        onSuccess: () => {
-          setIsFollowing(true);
-          setFollowerCount((prev) => prev + 1);
-        },
-        onError: (error) => {
-          console.error('팔로우 오류:', error);
-        },
-      },
-    );
-  };
-
-  const handleUnfollow = () => {
-    if (!uid) return;
-    unfollowMutate(
-      { uid: Array.isArray(uid) ? uid[0] : uid, nickname: userInfo?.nickname || '' },
-      {
-        onSuccess: () => {
-          setIsFollowing(false);
-          setFollowerCount((prev) => Math.max(prev - 1, 0));
-        },
-        onError: (error) => {
-          console.error('언팔로우 오류:', error);
-        },
-      },
-    );
   };
 
   if (!userInfo) return <p>로딩 중...</p>;
@@ -182,7 +67,9 @@ const RootLayout = ({ children }: RootLayoutProps) => {
         <div className="flex flex-row mb-6">
           <p
             className={`text-xl font-black ml-4 mt-6 p-2 cursor-pointer ${
-              isCommunityActive ? 'text-black border-b-4 border-black' : 'text-gray-400'
+              isCommunityActive
+                ? 'text-black border-b-4 border-black'
+                : 'text-gray-400'
             }`}
             onClick={() => handleTabClick(`/channel/${uid}`)}
           >
@@ -190,7 +77,9 @@ const RootLayout = ({ children }: RootLayoutProps) => {
           </p>
           <p
             className={`text-xl font-black ml-4 mt-6 p-2 cursor-pointer ${
-              activeSegment === 'follow' ? 'text-black border-b-4 border-black' : 'text-gray-400'
+              activeSegment === 'follow'
+                ? 'text-black border-b-4 border-black'
+                : 'text-gray-400'
             }`}
             onClick={() => handleTabClick(`/channel/${uid}/follow`)}
           >
